# Reprecord.Backend/openApi/openapi.yaml
openapi: 3.0.0
info:
    title: Gym Workout Tracker API
    version: 1.0.0
    description: API for managing gym workouts, exercise logs, and user progress.
servers:
    - url: http://localhost:3000/api
      description: Local development server
    - url: http://localhost:3001/api
      description: Local dev server
...
paths:
...
    /users/me:
        get:
            tags:
                - Users
            operationId: getCurrentUserProfile
            x-eov-operation-handler: controllers/UserController
            summary: Get the current user's profile
            responses:
                "200":
                    description: User profile
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
        put:
            tags:
                - Users
            operationId: updateCurrentUserProfile
            x-eov-operation-handler: controllers/UserController
            summary: Update the current user's profile
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/User"
            responses:
                "200":
                    description: User updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/User"
    /workout-plans:
        get:
            tags:
                - Workout Plans
            operationId: getAllWorkoutPlans
            x-eov-operation-handler: controllers/WorkoutPlanController
            summary: Get all workout plans for the logged-in user
            responses:
                "200":
                    description: List of workout plans
                    content:
                        application/json:
                            schema:
                                type: array
                                items:
                                    $ref: "#/components/schemas/WorkoutPlan"
        post:
            tags:
                - Workout Plans
            operationId: createWorkoutPlan
            x-eov-operation-handler: controllers/WorkoutPlanController
            summary: Create a new workout plan
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkoutPlan"
            responses:
                "201":
                    description: Workout plan created
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkoutPlan"
    /workout-plans/{id}:
        get:
            tags:
                - Workout Plans
            operationId: getWorkoutPlanById
            x-eov-operation-handler: controllers/WorkoutPlanController
            summary: Get a specific workout plan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "200":
                    description: Workout plan details
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkoutPlan"
        put:
            tags:
                - Workout Plans
            operationId: updateWorkoutPlanById
            x-eov-operation-handler: controllers/WorkoutPlanController
            summary: Update a specific workout plan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            requestBody:
                content:
                    application/json:
                        schema:
                            $ref: "#/components/schemas/WorkoutPlan"
            responses:
                "200":
                    description: Workout plan updated
                    content:
                        application/json:
                            schema:
                                $ref: "#/components/schemas/WorkoutPlan"
        delete:
            tags:
                - Workout Plans
            operationId: deleteWorkoutPlanById
            x-eov-operation-handler: controllers/WorkoutPlanController
            summary: Delete a specific workout plan
            parameters:
                - name: id
                  in: path
                  required: true
                  schema:
                      type: string
                      format: uuid
            responses:
                "204":
                    description: Workout plan deleted
...
components:
    schemas:
        User:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                username:
                    type: string
                email:
                    type: string
                password:
                    type: string
        WorkoutPlan:
            type: object
            properties:
                id:
                    type: string
                    format: uuid
                userId:
                    type: string
                    format: uuid
                name:
                    type: string
                createdAt:
                    type: string
                    format: date-time
                updatedAt:
                    type: string
                    format: date-time
...
    securitySchemes:
        bearerAuth:
            type: http
            scheme: bearer
            bearerFormat: JWT